+++ test/east-human/case16_instance_member.cpp
@@ -7,21 +7,21 @@ namespace east_view /* source: test/py/case16_instance_member.py */ {
 // [5:0] class original=Point
 struct Point {
 // [6:4] function original=__init__
-None __init__(Point self /* mutable */, int x /* readonly */, int y /* readonly */) {
+None __init__(Point self /* mutable */, int64 x /* readonly */, int64 y /* readonly */) {
             // [7:8]
-            self.x /* type=int, borrow=value */ = x /* type=int, borrow=readonly_ref */;
+            self.x /* type=int64, borrow=value */ = x /* type=int64, borrow=readonly_ref */;
             // [8:8]
-            int self.y /* type=int, borrow=value */ = y /* type=int, borrow=readonly_ref */;
+            int64 self.y /* type=int64, borrow=value */ = y /* type=int64, borrow=readonly_ref */;
         }
 // [10:4] function original=total
-int total(Point self /* readonly */) {
+int64 total(Point self /* readonly */) {
             // [11:8]
-            return self.x + self.y /* type=int, borrow=value */;
+            return self.x + self.y /* type=int64, borrow=value */;
         }
     };
 
     // main guard body
-    int __east_main_guard() {
+    int64 __east_main_guard() {
         // [15:4]
         Point p /* type=Point, borrow=value */ = Point(2, 5) /* type=Point, borrow=value */;
         // [16:4]
